Add ability to specify a specific pair of controller and action
With the actual behavior it is not possible to define specific actions to match
on a controller if we send the option as `[['controller', 'controller_b'],
['action', 'action_b']]` this will match `controller#action`,
`controller#action_b`, `controller_b#action` and `controller_b#action_b`.

Add the ability to match a specific controller and action using the following
syntax `[controller: :action, controller_b: :action_b]` this will match
`controller#action` and `controller_b#action_b` and does not match
`controller#action_b` and `controller_b#action`.